+++ imdb/__init__.py
@@ -25,7 +25,7 @@ from __future__ import absolute_import, division, print_function, unicode_litera
 
 __all__ = ['IMDb', 'IMDbError', 'Movie', 'Person', 'Character', 'Company',
            'available_access_systems']
-__version__ = VERSION = '6.7dev20181218'
+__version__ = VERSION = '6.7dev2018124'
 
 
 import logging